chore: separate graphql and rest controllers
